updates
scrapped my old idea and decided to make a json file (food banks) becayes the way it is formatted makes it easier to print out all the info. Updated the directory funcs so it works, but the json info is empty for now. tried making a thing where the user can bookmark the place to a file they named which is stored in bookmarks.txt (idk if it works cause I didn't test it yet).